requirements.txt
streamlit==1.28.1
openai==1.3.5
pandas==2.1.3
sqlalchemy==2.0.23
python-dotenv==1.0.0
plotly==5.17.0
psycopg2-binary==2.9.9
mysql-connector-python==8.2.0
sqlite3
